keyword_classfiy_new/pyproject.toml
[project]
name = "keyword-classfiy"
version = "0.2.7"
description = "关键词分类工具，支持复杂的分词规则匹配"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "lark-parser>=0.12.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0.3",
    "pathos>=0.3.3",
    "pyinstaller>=6.12.0",
    "toml>=0.10.2",
]

